Don't leak drmModePageFlip events
Tidying up the hack a little to not request an event back from drmModePageFlip
if we know we can't handle it. Otherwise we ended up running out of memory in
the kernel because all the events got leaked.

As it's now two changes, added a quick #if to handle both cases.