Move framebuffer tracking into omap_bo
A DRM framebuffer is immutable so it makes more sense for an omap_bo
to track its DRM framebuffer.

Previously we ended up deleting a framebuffer which was being scanned
out, ending up in nothing being scanned out. Now the framebuffer is
only deleted when the omap_bo is deleted, which is only deleted when
the pixmap is deleted.

Still remaining is to never explicitly delete pOMAP->scanout, as it
will get deleted in the right place when the pixmap wrapping it is
finally deleted (it's refcounted). Also, pOMAP->scanout needs updating
whenever a page flip or exchange happens.

Conflicts:

	src/drmmode_display.c