Fix transition between non-flippable and flippable
Before this patch it was not possible for a window to be moved to
the origin and have it begin page flipping. Because the size didn't
change, the DRI2 buffers would never get re-allocated and the buffers
would remain without a DRM fb and therefore unable to be flipped to.

This change forces DRI2 buffers to be re-allocated whenever a
drawable transitions between being flippable and non-flippable, and
vice versa. This both allows previously non-flippable windows to
start flipping, and saves possibly scarce scanout-capable memory
when a flippable window becomes non-flippable again.

As a somewhat unrelated change, this patch also removes the
possibility of exchanging pixmaps to effect a swap. This is because
it crashed whenever we hit this case, and we don't understand enough
what is meant by exchange to be able to implement it properly.