Add DRM driver abstraction
This adds an abstraction interface in src/drmmode_driver.h for the
driver-specific values (and in future functions) that we need. It
also adds documentation and build system support (--with-drmmode)
to choose your supported driver.

Initially, the dumb allocation flags and page flip events flags
are abstracted with support for pl111 and eynos.

A template implementation is also provided for ease of adding new
driver support.

Change-Id: I98890ddab6b907c3007b72a662ecd7ee6caa4402