Add InitFromFBDev xorg.conf option
Adds a new option that can be specified in xorg.conf - InitFromFBDev.
With this set to a valid fbdev device file the contents of the DRM
scanout buffer will be initialized by bliting from the fbdev device.

This only makes sense when the Xorg -background none parameter is set
on startup, otherwise the background would be overwritten.

The primary use for this feature is to enable a smooth transition
from fbdev to X.

This commit is based on a set of commits to the armsoc fork at:
https://chromium.googlesource.com/chromiumos%2Fthird_party%2Fxf86-video-armsoc
Commits:
  a815781, 72af3eb, 0d7e384, 8be5d4f, 2ada542, 2e46d97, 2e46d97

NB: This implementation uses pixman to do the blit

Change-Id: I227c120954a2a3f768d67bc04c00cbcd30917b64