Ensure the change of window flippablity before the first ScheduleSwap is always caught
The flippability of a window could changes between
ARMSOCDRI2CreateBuffer and the first call of
ARMSOCDRI2ScheduleSwap. This change is not picked
up by ScheduleSwap and hence the buffer of the window
is not forced to update. This change fixes that behavior.

This fixes situations where:
- Window gets mapped
- Window resize nonfs->fs (or vice versa)
- Window gets unredirected
- Fullscreen window gets unobscured
between eglCreateWindowSurface and eglSwapBuffers.

Change-Id: Ica82814638d6fbd91f5db984692d5c52ac3f910e