Change custom GEM creation flags for pl111.
GEM flags in the pl111 DRM driver have been changed
so that we can specify the backing storage (SHM/DMA)
and the kind of user space mapping (non cached, cached,
writecombine).

Support for cached mappings for SHM objects is not
stable yet, so keep on using non cached mappings
for now.

Change-Id: I0085d01a37f142d8ad196ecb0d1624ff4f15a1ef