Create UMP handle for each GEM allocation
Use a new ioctl in exynos-drm to allocate memory and also get a UMP
ID along with it. In the Mali world, UMP IDs are passed in the DRI2
buffer name in order to share GL buffers between client and server.

Kernel patch is https://gist.github.com/dsd/8fe8b84fe53e2b0688f5

Thanks to Daniel Stone and Sjoerd Simons for basically spelling
out this patch in an IRC discussion.