DRI2 CreateBuffers: suppress buffer reuse
When returning a buffer from CreateBuffers, the DRI2 code in the
X server caches it so that it can return it for future DRI2GetBuffers
queries too.

Until we get a better understanding of the precise event sequence
and behaviour that Mali blobs use, I can't fully explain why this is
needed. It is to do with understanding which buffer Mali is using for
front and back. Without this in place, we often send the wrong buffer
over scanout.

fbturbo does the same thing, although for slightly different reasons.
https://github.com/ssvb/xf86-video-fbturbo/wiki/Mali400-DRI2