Make ScheduleSwap block on vblank
ScheduleSwap is designed to be asynchronous, scheduling a page flip
for a future vblank, and returning immediately. A DRI2 "swap completed"
event is generated later to indicate when the old front buffer is available
for use as the next back buffer.

While my grasp of the DRI2 event sequence and handling for the mali blobs
is not totally complete, I think I have deduced that Mali gets this totally
wrong. I think it assumes that ScheduleSwap completes synchronously and the
old front buffer is available for reuse as soon as this function returns.
This causes a visual flicker:
 1. Mali draws to back buffer
 2. ScheduleSwap is called, that buffer will now be sent to the scanout
    starting from next vblank.
 3. ScheduleSwap returns without blocking. Mali immediately reuses and
    modifies the old frontbuffer, which is still being displayed on the
    screen. Screen goes white (momentarily, until we hit the next step).
 4. vblank happens and the page flip completes, now the output image is
    good.

Making ScheduleSwap block on the vblank avoids this. Now when ScheduleSwap
returns, the front buffer is truly available for reuse.
As an initial hack, I do this with usleep(). Actually usleep() seems to
not really sleep at all, so we run a busy-loop. Can't go much below 200
as the magic number here without the flicker returning.

Before making this discovery, I was battling with another oddity and
had to introduce another "fix" in the same place. Mali sometimes
requests 2 page flips in quick succession, without waiting for the first
one to finish. As the first one has not finished, exchangebufs() has not
been called in time for the 2nd call, but by checksumming the buffers
at various points I confirmed that Mali really does behave like this:

- Buffer A modified
- SwapBuffers called (swaps A with old front buffer B)
- Page flip completes on vblank soon after
- Buffer B modified, A is unmodified
- SwapBuffers called (swaps B with A), doesn't wait for vblank
- Buffer A modified, B is unmodified
- SwapBuffers called again

Before this commit, the final SwapBuffers call would have tried to
swap B with A again. However, in this case, buffer A is the back buffer,
confirmed by the fact that it just got modified (verified with checksums).
To work around this, exchangebufs() should be called immediately after
the page flip is requested - don't wait for it to complete.

However, the exchangebufs "fix" is now made irrelevant by the sleep.
The sleep avoids the above sequence of events, by ensuring that
the page flip has actually completed before returning from ScheduleSwap.

Thanks to Daniel Stone for helping to think through this situation.