Synchronize texture access with umplock
Via DRI2 CreateBuffers we give the underlying pixmap memory addresses
to Mali. Both Mali and X then access the pixmaps, so we have a
synchronization problem to solve.

Fortunately, libMali can be built with "umplock" functionality,
which basically (hopefully) means that Mali informs the kernel when
it is using a specific texture. We can then use umplock here too,
in order to guarantee synchronized access.