Scanout buffer reference counting fixes
When changing resolution we would hit assertion failures, attemping to
work with freed buffers.

Improve the reference counting so that a reference is added on the
scanout buffer at server startup, transferred to the new scanout
buffer when it gets changed, and freed on shutdown.

When ModifyPixmapHeader also binds the scanout texture to a pixmap,
we have to create a reference there too, because we unref it in
DestroyPixmap.

Fixes crashes and leaks when changing resolution. For testing, I
ran the following scenarios, ensuring that there were no crashes
or memory leaks:

1. Start raw X session then start and stop es2gears many times.
2. Start raw X session, start and stop mutter many times.
3. Start raw X session, start mutter, change the resolution many times.
4. Start raw X session, start gnome-shell, change the resolution many times.