Revert "Move wakeup handler to per-fd not per-Screen"
This reverts commit 04af999.

This does not work quite right as because the server calls
InitBlockAndWakeupHandlers() for each server generation, which wipes
out any event handlers that were installed with the changes in this
patch.

Fixes the ability to run:
  X &
  mutter &
  xterm &
  killall mutter xterm
  mutter &
  xterm &

etc.

Without this patch, mutter does not fully draw the xterm window on
screen.