Check for encoder before checking for overscan compensation
When X is started without a display, and then you connect a display
later, we'll arrive in this function with no encoder available
(X is seeing what display modes are available before deciding whether
to enable the display or not). The function is then called again later
with an encoder available.

Fix a segfault when there is no encoder.