Fix refcounting when exchanging buffers
When exchanging buffers the old bo associated to a buffer might not
be the same as the new one (e.g. after a resize). So save the old
buffer pointers in ScheduleSwap to match them with the unreference calls
in SwapComplete

This match the approach taken in the upstream armsoc driver.