drmmode_display: Stop animated cursors from flickering on meson
meson is an atomic driver, meaning that it attempts to set the
cursor image with drmModeSetCursor wait until vblank before
completing. For historical reasons, ARMSOC hides the cursor
before updating the pixels in the buffer, which means that we
hide, wait for vblank, show new image, wait for vblank, leading
to awful flickering of the cursor. This is especially visible
during animated cursors like the spinning beachball when opening
an app.

Remove this code. ... and then realize that Xorg does the same
thing in its HW cursor manipulating code, meaning all that code
before was simply dead. Pass a flag to Xorg to tell it to stop
that. We make great APIs.