armsoc_driver: Make one depth-32 lie match our other one
We lie and say that the root window is depth-32 in the miSetVisualTypes
call below, but when we create the scanout buffer, we actually say it's
depth-24, which isn't right.

When we page flip, the kernel actually double-checks to make sure that
the pixel formats match up, and if not, it returns an error. And our
DDX is not smart enough to recover from a failing page flip right now,
unfortunately.

Through a dumb trick of luck, the Mali driver used to lie a lot about
the depths it was using, and these two idiosyncrasies matched up.

So, we used to set an ARGB8888 scanout buffer and pretend it was
depth-24. Mali told us to page flip an ARGB8888 buffer, which it also
said was depth-24 (a bug), and it worked fine.

When Mali was upgraded to r5p1, they fixed this depth-24 bug, and it now
accurately says it's depth-32, which causes the kernel to complain.

Just tell the kernel our scanout buffer is depth-32. I don't know if
it's right, wrong, or insane, but it prevents the kernel from
complaining a lot.