Add a hack to force the alpha channel of the scanout buffer
The X server constantly thinks it can ignore the alpha channel if it's
not included in the depth. Unfortunately, including the alpha channel in
the depth tends to have consequences, so we need a depth 24 root window.

This turns on a hack path that makes sure not to overwrite the alpha
channel of the scanout buffer when painting with the X server.