Use the bpp always when reporting the depth to the kernel
Otherwise, when we go to flip, the depth of the buffer we want to flip
to might not actually match the depth of the scanout buffer itself. The
kernel considers this an error condition and doesn't let us page flip.

This fixes running egl-example fullscreen on a raw X11 server.