armsoc_exa: Don't wait when we're interrupted by a signal
The umplock driver uses down_interruptible for its semaphore when
protecting its internals. The X server is very signal-heavy: every mouse
move or key press makes us wake up with SIGIO and do a poll. This can
interrupt our syscall with the kernel, and we then sleep for 2ms, which
can have disasterous effects when stacked, from general unresponsiveness
to graphics corruption issues.

Perhaps in the future we'll use down_killable instead of
down_interruptiblle in the umplock driver. For now, just don't usleep
when we get EINTR.