Implement shadow buffer
When using xrandr to switch between two outputs on a CRTC, for some
reason we end up in codepaths that require shadow buffers.

Implement this, to avoid X crashing. When a shadow buffer is created
by this codepath, it is destroyed immediately after.

I confirmed that this code does not trigger during normal usage,
nor when displays are switched by mutter.

[endlessm/eos-shell#5893]