Ignore cache control failure
There is a corner case that can result in us attempting to
change the domain of a bo that was allocated as scanout.

It's easy to reproduce by running a fullscreen unredirected chromium and
pressing the volume keys (causing an OSD to appear), or to alt-tab to
another window. It usually reproduces after a few attempts.

It's not exactly clear why this is happening, but for now, just ignore
cache control ioctl failure, like Mali does.

https://phabricator.endlessm.com/T6160