Force PIXMAP_SCANOUT when the pixmap has the same size as the screen
If the window is not mapped when it is created, the backing pixmap is
created as BACKING_PIXMAP. When later during execution it is mapped
again, a new framebuffer will be created for that BO.

The problem with this flow is that with the latest modifications on
the meson platform the non-scanout BOs are memory scattered. This means
that we cannot allocate a framebuffer from them.

With this patch the pixmap is always allocated as ARMSOC_BO_SCANOUT
when the drawable size has the same size as the screen.

Signed-off-by: Carlo Caione <carlo@endlessm.com>