Conditionally reallocate buffers in ReuseBufferNotify
CreateBuffers is sometimes called for a non-flippable pixmap which
later becomes flippable. For example, when a pixmap is not yet mapped,
or when creating a window pixmap that later becomes fullscreen.

In this case we end up allocating backing pixmap memory, which we later
attempt to use as scanout. With our recent kernel changes to scatter
allocations for pixmaps, that's no longer permitted.

When a drawable goes through a state transition like this, we do have
the opportunity to detect and correct the situation in ReuseBufferNotify.
Here we can extend existing infrastructure to completely reallocate the
backing object.

We also do the same in the reverse case: when a scanout buffer is
no longer being used for scanout, we reallocate it as a pixmap buffer,
saving on precious scanout memory.

Rather than duplicating a lot of buffer create/destroy code, I factored
out the common parts into new functions, CreateBufferResources and
DestroyBufferResources.

[endlessm/eos-shell#6348]